fixed ：ESM 下 prototype 为 undefined、logical-model 挂起超时、以及 domService.requestRead 未 mock
