+++ web/projects/portal/src/app/composer/gui/vs/objects/selection/drag-border-type.ts
@@ -1,0 +1,30 @@
+/*
+ * This file is part of StyleBI.
+ * Copyright (C) 2026  InetSoft Technology
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+/**
+ * Shared drag-border constants for selection-container children and
+ * EditableObjectContainer. Kept in a separate module so EditableObjectContainer
+ * does not statically import composer-selection-container-children.component
+ * (that cycle left ComposerSelectionContainerChildren undefined under CI ESM order).
+ */
+export enum DragBorderType {
+   NONE = 0, // "none"
+   ABOVE = 1, // "border-top"
+   ALL = 2, // "border"
+   BELOW = 3, // "border-bottom"
+}
